Move Listing stories into a directory, move mock data into DataMocks.js
